requirements.txt
pyyaml
tomli
tomlkit
rich
typer
appdirs
loguru
faker
aiofiles
aiohttp
aiohttp-socks
python-dateutil
ddddocr>=1.5.5
embypyfix
pyrogram
telethon
tgcrypto
pillow>10.0.0
thefuzz[speedup]
schema
aiocache
flask==3.0.3
flask-socketio==5.3.6
flask-login==0.6.3
simple-websocket
eventlet
cryptography
werkzeug==3.0.3
cachetools
emoji
numpy<2.0.0
onnxruntime==1.14.0
httpx[http2]
socksio
